# control-plane/src/main/resources/static/openapi/token-operations-v1.yaml
openapi: 3.1.0
info:
  title: Digital Banking Durable Operations API
  version: 1.0.0
  description: >-
    Records durable mint and burn operation requests and chain-neutral transfer
    acceptance. HTTP 202 means the local acceptance transaction committed; it does
    not perform signing, submission, minting, burning, bank effects, reconciliation,
    settlement, or any external side effect.
paths:
  /v1/token-operations/mints:
    post:
      operationId: acceptMintOperation
      summary: Durably accept a mint operation request
      description: >-
        A replay with the same participant scope, key, and canonical request
        returns HTTP 202 and the same operation without creating another event.
      security:
        - identityAdapter: []
      x-required-authority: token:mint
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AcceptanceRequest'
            example:
              contractVersion: 1
              assetId: REFERENCE_ASSET
              unitId: REFERENCE_UNIT
              unitVersion: 3
              quantity: '12.34'
              businessCorrelation: corr-001
      responses:
        '202':
          $ref: '#/components/responses/Accepted'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '422':
          $ref: '#/components/responses/Unprocessable'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /v1/token-operations/burns:
    post:
      operationId: acceptBurnOperation
      summary: Durably accept a burn operation request
      description: >-
        A replay with the same participant scope, key, and canonical request
        returns HTTP 202 and the same operation without creating another event.
      security:
        - identityAdapter: []
      x-required-authority: token:burn
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AcceptanceRequest'
      responses:
        '202':
          $ref: '#/components/responses/Accepted'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '422':
          $ref: '#/components/responses/Unprocessable'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /v1/token-operations/{operationId}:
    get:
      operationId: getTokenOperation
      summary: Read a participant-scoped durable operation
      security:
        - identityAdapter: []
      x-required-authority: token:read
      parameters:
        - name: operationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Durable operation state and histories
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenOperation'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /v1/transfers:
    post:
      operationId: acceptTransfer
      summary: Durably accept a chain-neutral transfer plan
      description: >-
        Commits one parent transfer, five ordered planned effects, immutable
        server-resolved wallet context, initial history, idempotency, and one
        TransferAccepted outbox event. It performs no bank or blockchain effect.
      security:
        - identityAdapter: []
      x-required-authority: transfer:create
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransferAcceptanceRequest'
            example:
              amount: '12.34'
              currency: USD
              sourceBankAccountReference: synthetic-bank:source-001
              destinationBankAccountReference: synthetic-bank:destination-001
              settlementNetwork: ETHEREUM
      responses:
        '202':
          $ref: '#/components/responses/TransferAccepted'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '422':
          $ref: '#/components/responses/Unprocessable'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /v1/transfers/{transferId}:
    get:
      operationId: getTransfer
      summary: Read a participant-scoped durable transfer plan
      security:
        - identityAdapter: []
      x-required-authority: transfer:read
      parameters:
        - name: transferId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Durable parent state, five effect states, and separate finalities
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transfer'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/TransferNotFound'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...
components:
  securitySchemes:
    identityAdapter:
      type: http
      scheme: bearer
      description: >-
        Identity-adapter boundary only. This reference phase configures no
        decoder, issuer, JWK, static token, password, or local user.
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: >-
        Opaque caller key scoped by tenant, participant, resource, and operation
        kind. It contains 1 to 128 visible US-ASCII characters, is hashed before
        storage, and is never returned.
      schema:
        type: string
        minLength: 1
        maxLength: 128
        pattern: '^[!-~]{1,128}$'
  responses:
    Accepted:
      description: >-
        The local operation, audit, finality, idempotency, and pending outbox
        records committed atomically. Processing has not started.
      headers:
        Location:
          required: true
          description: Stable participant-scoped operation resource
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/TokenOperation'
          example:
            operationId: 8c5f0a89-9317-4f30-a7ca-f18a395293ce
            kind: MINT
            asset:
              assetId: REFERENCE_ASSET
              unitId: REFERENCE_UNIT
              unitVersion: 3
              scale: 2
            quantity: '12.34'
            state: REQUESTED
            version: 0
            createdAt: '2026-07-16T23:00:00.123456Z'
            updatedAt: '2026-07-16T23:00:00.123456Z'
            businessCorrelation: corr-001
            attempts: []
            transitions: []
            finalities:
              blockchain:
                - status: NOT_ASSESSED
                  updatedAt: '1970-01-01T00:00:00Z'
                  evidenceReferences: []
              legal:
                - status: NOT_ASSESSED
                  updatedAt: '1970-01-01T00:00:00Z'
                  evidenceReferences: []
              customerVisible:
                - status: NOT_ASSESSED
                  updatedAt: '1970-01-01T00:00:00Z'
                  evidenceReferences: []
              accounting:
                - status: NOT_ASSESSED
                  updatedAt: '1970-01-01T00:00:00Z'
                  evidenceReferences: []
            evidenceReferences:
              - participant:acceptance:48ff09c2-0a2e-40c4-a122-190a66fc7444
    TransferAccepted:
      description: >-
        The local parent, five planned effects, histories, idempotency, and
        pending outbox event committed atomically. No financial effect started.
      headers:
        Location:
          required: true
          description: Stable participant-scoped transfer resource
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Transfer'
          example:
            transferId: 00000000-0000-0000-0000-000000000100
            amount: '12.34'
            currency: USD
            sourceBankAccountReference: synthetic-bank:source-001
            destinationBankAccountReference: synthetic-bank:destination-001
            settlementNetwork: ETHEREUM
            status: ACCEPTED
            version: 0
            createdAt: '2026-07-17T18:00:00.123456Z'
            updatedAt: '2026-07-17T18:00:00.123456Z'
            effects:
              - effectId: 00000000-0000-0000-0000-000000000101
                sequence: 1
                kind: BANK_WITHDRAWAL
                status: PLANNED
                attemptCount: 0
                evidenceReferences: []
              - effectId: 00000000-0000-0000-0000-000000000102
                sequence: 2
                kind: TOKEN_MINT
                status: PLANNED
                attemptCount: 0
                evidenceReferences: []
              - effectId: 00000000-0000-0000-0000-000000000103
                sequence: 3
                kind: TOKEN_TRANSFER
                status: PLANNED
                attemptCount: 0
                evidenceReferences: []
              - effectId: 00000000-0000-0000-0000-000000000104
                sequence: 4
                kind: TOKEN_BURN
                status: PLANNED
                attemptCount: 0
                evidenceReferences: []
              - effectId: 00000000-0000-0000-0000-000000000105
                sequence: 5
                kind: BANK_DEPOSIT
                status: PLANNED
                attemptCount: 0
                evidenceReferences: []
            finalities:
              blockchain:
                - status: NOT_ASSESSED
                  updatedAt: '1970-01-01T00:00:00Z'
                  evidenceReferences: []
              legal:
                - status: NOT_ASSESSED
                  updatedAt: '1970-01-01T00:00:00Z'
                  evidenceReferences: []
              customerVisible:
                - status: NOT_ASSESSED
                  updatedAt: '1970-01-01T00:00:00Z'
                  evidenceReferences: []
              accounting:
                - status: NOT_ASSESSED
                  updatedAt: '1970-01-01T00:00:00Z'
                  evidenceReferences: []
            evidenceReferences:
              - participant:transfer:acceptance:00000000-0000-0000-0000-000000000100
    BadRequest:
      description: Malformed request or invalid bounded input
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: urn:digital-banking:problem:invalid-request
            title: Invalid request
            status: 400
            detail: The request did not satisfy the API contract.
            instance: urn:digital-banking:problem-instance:undisclosed
    Unauthenticated:
      description: Authenticated participant identity is required
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: urn:digital-banking:problem:unauthenticated
            title: Authentication required
            status: 401
            detail: An authenticated participant identity is required.
    Unauthorized:
      description: Required operation authority is absent
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: urn:digital-banking:problem:unauthorized
            title: Access denied
            status: 403
            detail: The authenticated participant is not authorized for this operation.
    NotFound:
      description: >-
        Operation is unknown or outside the authenticated participant scope;
        these cases are intentionally indistinguishable.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: urn:digital-banking:problem:operation-not-found
            title: Operation not found
            status: 404
            detail: The operation was not found.
            instance: urn:digital-banking:problem-instance:undisclosed
    TransferNotFound:
      description: >-
        Transfer is unknown or outside the authenticated participant scope;
        these cases are intentionally indistinguishable.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: urn:digital-banking:problem:transfer-not-found
            title: Transfer not found
            status: 404
            detail: The transfer was not found.
            instance: urn:digital-banking:problem-instance:undisclosed
    Conflict:
      description: The scoped idempotency key names a different canonical request
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: urn:digital-banking:problem:idempotency-conflict
            title: Idempotency conflict
            status: 409
            detail: The idempotency key is already bound to a different request.
            instance: urn:digital-banking:problem-instance:undisclosed
    UnsupportedMediaType:
      description: Request content is not application/json
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: urn:digital-banking:problem:unsupported-media-type
            title: Unsupported media type
            status: 415
            detail: Use application/json for this request.
            instance: urn:digital-banking:problem-instance:undisclosed
    Unprocessable:
      description: Unsupported request contract, server-owned asset unit, or transfer route
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: urn:digital-banking:problem:unprocessable-request
            title: Unprocessable request
            status: 422
            detail: The request uses an unsupported contract, asset unit, or transfer route.
            instance: urn:digital-banking:problem-instance:undisclosed
    ServiceUnavailable:
      description: Durable acceptance or read-back is unavailable
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: urn:digital-banking:problem:service-unavailable
            title: Service unavailable
            status: 503
            detail: Durable acceptance is temporarily unavailable.
            instance: urn:digital-banking:problem-instance:undisclosed
    InternalError:
      description: Unexpected internal invariant or mapping failure
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: urn:digital-banking:problem:internal-error
            title: Internal server error
            status: 500
            detail: An unexpected internal error occurred.
            instance: urn:digital-banking:problem-instance:undisclosed
  schemas:
    CanonicalQuantity:
      type: string
      minLength: 1
      maxLength: 512
      pattern: '^(0|[1-9][0-9]*)(\.[0-9]*[1-9])?$'
      description: Positive canonical decimal text; never a JSON number
      examples:
        - '12.34'
    AcceptanceRequest:
      type: object
      additionalProperties: false
      required:
        - contractVersion
        - assetId
        - unitId
        - unitVersion
        - quantity
        - businessCorrelation
      properties:
        contractVersion:
          type: integer
          const: 1
        assetId:
          $ref: '#/components/schemas/SafeIdentifier'
        unitId:
          $ref: '#/components/schemas/SafeIdentifier'
        unitVersion:
          type: integer
          minimum: 1
        quantity:
          $ref: '#/components/schemas/CanonicalQuantity'
        businessCorrelation:
          type: string
          minLength: 1
          maxLength: 128
    TransferAcceptanceRequest:
      type: object
      additionalProperties: false
      required:
        - amount
        - currency
        - sourceBankAccountReference
        - destinationBankAccountReference
      properties:
        amount:
          $ref: '#/components/schemas/CanonicalQuantity'
        currency:
          type: string
          pattern: '^[A-Z][A-Z0-9]{2,11}$'
        sourceBankAccountReference:
          $ref: '#/components/schemas/SyntheticBankAccountReference'
        destinationBankAccountReference:
          $ref: '#/components/schemas/SyntheticBankAccountReference'
        settlementNetwork:
          type: string
          enum: [ETHEREUM, SOLANA]
          description: Optional logical future-adapter route; defaults locally when absent
    SyntheticBankAccountReference:
      type: string
      pattern: '^synthetic-bank:[A-Za-z0-9][A-Za-z0-9._:-]{0,111}$'
      description: Opaque synthetic reference; never real account or routing data
    SafeIdentifier:
      type: string
      pattern: '^[A-Za-z0-9][A-Za-z0-9._:-]{0,63}$'
    TokenOperation:
      type: object
      additionalProperties: false
      required:
        - operationId
        - kind
        - asset
        - quantity
        - state
        - version
        - createdAt
        - updatedAt
        - businessCorrelation
        - attempts
        - transitions
        - finalities
        - evidenceReferences
      properties:
        operationId:
          type: string
          format: uuid
        kind:
          type: string
          enum: [MINT, BURN]
        asset:
          $ref: '#/components/schemas/AssetUnit'
        quantity:
          $ref: '#/components/schemas/CanonicalQuantity'
        state:
          type: string
          description: Durable lifecycle state; acceptance begins at REQUESTED
        version:
          type: integer
          format: int64
          minimum: 0
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        businessCorrelation:
          type: string
        attempts:
          type: array
          items:
            $ref: '#/components/schemas/Attempt'
        transitions:
          type: array
          items:
            $ref: '#/components/schemas/Transition'
        finalities:
          $ref: '#/components/schemas/FinalityHistories'
        evidenceReferences:
          type: array
          description: Participant-safe opaque references only; internal evidence is omitted
          items:
            type: string
    AssetUnit:
      type: object
      additionalProperties: false
      required: [assetId, unitId, unitVersion, scale]
      properties:
        assetId:
          $ref: '#/components/schemas/SafeIdentifier'
        unitId:
          $ref: '#/components/schemas/SafeIdentifier'
        unitVersion:
          type: integer
          minimum: 1
        scale:
          type: integer
          minimum: 0
          maximum: 255
    Attempt:
      type: object
      additionalProperties: false
      required: [attemptId, predecessor, createdAt]
      properties:
        attemptId:
          type: string
          format: uuid
        predecessor:
          type: [string, 'null']
          format: uuid
        createdAt:
          type: string
          format: date-time
    Transition:
      type: object
      additionalProperties: false
      required: [version, from, to, occurredAt, evidenceReferences]
      properties:
        version:
          type: integer
          format: int64
        from:
          type: string
        to:
          type: string
        occurredAt:
          type: string
          format: date-time
        evidenceReferences:
          type: array
          description: Participant-safe opaque references only
          items:
            type: string
    FinalityHistories:
      type: object
      additionalProperties: false
      required: [blockchain, legal, customerVisible, accounting]
      properties:
        blockchain:
          $ref: '#/components/schemas/FinalityHistory'
        legal:
          $ref: '#/components/schemas/FinalityHistory'
        customerVisible:
          $ref: '#/components/schemas/FinalityHistory'
        accounting:
          $ref: '#/components/schemas/FinalityHistory'
...
    Transfer:
      type: object
      additionalProperties: false
      required:
        - transferId
        - amount
        - currency
        - sourceBankAccountReference
        - destinationBankAccountReference
        - settlementNetwork
        - status
        - version
        - createdAt
        - updatedAt
        - effects
        - finalities
        - evidenceReferences
      properties:
        transferId:
          type: string
          format: uuid
        amount:
          $ref: '#/components/schemas/CanonicalQuantity'
        currency:
          type: string
        sourceBankAccountReference:
          $ref: '#/components/schemas/SyntheticBankAccountReference'
        destinationBankAccountReference:
          $ref: '#/components/schemas/SyntheticBankAccountReference'
        settlementNetwork:
          type: string
          enum: [ETHEREUM, SOLANA]
        status:
          type: string
          enum: [ACCEPTED, IN_PROGRESS, MANUAL_REVIEW, COMPENSATION_REQUIRED, EFFECTS_APPLIED]
          description: EFFECTS_APPLIED is not legal, accounting, customer, or settlement finality
        version:
          type: integer
          format: int64
          minimum: 0
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        effects:
          type: array
          minItems: 5
          maxItems: 5
          items:
            $ref: '#/components/schemas/TransferEffect'
        finalities:
          $ref: '#/components/schemas/TransferFinalities'
        evidenceReferences:
          type: array
          description: Participant-safe opaque references only
          items:
            type: string
    TransferEffect:
      type: object
      additionalProperties: false
      required: [effectId, sequence, kind, status, attemptCount, evidenceReferences]
      properties:
        effectId:
          type: string
          format: uuid
        sequence:
          type: integer
          minimum: 1
          maximum: 5
        kind:
          type: string
          enum: [BANK_WITHDRAWAL, TOKEN_MINT, TOKEN_TRANSFER, TOKEN_BURN, BANK_DEPOSIT]
        status:
          type: string
          enum: [PLANNED, PREPARED, ATTEMPT_PENDING, APPLIED, AMBIGUOUS, RETRYABLE_NO_EFFECT, TERMINAL_NO_EFFECT, MANUAL_REVIEW, COMPENSATION_REQUIRED]
        attemptCount:
          type: integer
          minimum: 0
        evidenceReferences:
          type: array
          description: Participant-safe opaque references only
          items:
            type: string
    TransferFinalities:
      type: object
      additionalProperties: false
      required: [blockchain, legal, customerVisible, accounting]
      properties:
        blockchain:
          $ref: '#/components/schemas/TransferFinalityHistory'
        legal:
          $ref: '#/components/schemas/TransferFinalityHistory'
        customerVisible:
          $ref: '#/components/schemas/TransferFinalityHistory'
        accounting:
          $ref: '#/components/schemas/TransferFinalityHistory'
...
    Problem:
      type: object
      required: [type, title, status, detail]
      properties:
        type:
          type: string
          format: uri
          enum:
            - urn:digital-banking:problem:invalid-request
            - urn:digital-banking:problem:unauthenticated
            - urn:digital-banking:problem:unauthorized
            - urn:digital-banking:problem:operation-not-found
            - urn:digital-banking:problem:transfer-not-found
            - urn:digital-banking:problem:idempotency-conflict
            - urn:digital-banking:problem:unsupported-media-type
            - urn:digital-banking:problem:unprocessable-request
            - urn:digital-banking:problem:internal-error
            - urn:digital-banking:problem:service-unavailable
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string
        instance:
          type: string
          format: uri